Feat: Services de produtos e services de registros, auto preenchimento de usuarios e autores de produtos

+++ src/main/java/com/example/drp/domain/user/User.java
@@ -42,7 +42,7 @@ public class User implements UserDetails {
     private Timestamp createdAt;
 
     @Column(name = "company_id")
-    private int companyId;
+    private long companyId;
 
     @Column(name = "role")
     private UserRole role;
